requirements.txt
requests==2.28.2
tiktoken==0.3.2
openai==0.27.2
python-dotenv==1.0.0